refactor: Javadoc links and fix cooldown type casting
Replaced plain URLs in Javadoc comments with HTML anchor tags for better formatting. Also corrected the type casting in CooldownTag to ensure cooldown values are set as long integers.

+++ core/src/main/java/com/sekwah/advancedportals/core/tags/CooldownTag.java
@@ -85,7 +85,7 @@ public class CooldownTag implements Tag.Activation, Tag.Creation {
                     playerDataServices.getPlayerData(player);
                 try {
                     playerData.setPortalCooldown(
-                        portal.getName(), Integer.parseInt(argData[0]) * 1000);
+                        portal.getName(), (long) Integer.parseInt(argData[0]) * 1000L);
                 } catch (NumberFormatException e) {
                     infoLogger.warning(
                         "Cooldown tag failed to set cooldown for portal: "
